adjusting routes and replace old inventory list page
